refactor: Restructure edit.py into edit/ package
Move generic functionality to shared modules for reuse by future create
enhancements:

- Add parse_monogram() and parse_object_ids() to core.py (generic)
- Move confirm_text_change() to editor.py (generic)
- Create yaml_utils.py with YAML parsing functions (generic)
- Create edit/ package with submodules:
  - core.py: Edit orchestrator class
  - validators.py: Board/column validation
  - formatters.py: Change display, partition suggestions
  - batch.py: Batch edit coordination
