+++ phabfive/edit/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""Edit package for phabfive."""
+
+from phabfive.edit.core import Edit
+
+__all__ = ["Edit"]
